requirements.txt
slack-sdk==3.19.5
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0